requirements/requirements.txt
colossalai>=0.4.0
mmengine>=0.10.3
pandas>=2.0.3
timm==0.9.16
rotary_embedding_torch==0.5.3
ftfy>=6.2.0 # for t5
diffusers==0.27.2 # for vae
accelerate==0.29.2 # for t5
av>=12.0.0 # for video loading
numpy<2.0.0
minio
fastapi[standard]
python-dotenv

# [gradio]
gradio>=4.26.0
spaces>=0.28.3
# [notebook]
ipykernel>=6.29.4
ipywidgets>=8.1.2
# [training]
wandb>=0.17.0
tensorboard>=2.14.0
pandarallel>=1.6.5
pyarrow>=16.1.0 # for parquet
# [dev]
pre-commit>=3.5.0
openai